added subscribe category period functionality

+++ application/lib/internship_organizer/period_manager/component/deleter.class.php
@@ -1,5 +1,8 @@
 <?php
 
+require_once Path :: get_application_path() . 'lib/internship_organizer/period_manager/component/browser.class.php';
+
+
 class InternshipOrganizerPeriodManagerDeleterComponent extends InternshipOrganizerPeriodManager
 {
 
@@ -26,39 +29,66 @@ class InternshipOrganizerPeriodManagerDeleterComponent extends InternshipOrganiz
             {
                 $period = $this->retrieve_period($id);
                 $parent_id = $this->get_parent();
-                if (! $period->delete())
+                
+                $status = true;
+                //if period has sub_periods, it isn't aloud to be deleted
+                
+
+                if ($period->has_children())
                 {
-                    $failures ++;
+                    $status = false;
+                    $message = 'PeriodNotDeleted-HasSubPeriods';
                 }
-                else
+                
+                $period_id = $period->get_id();
+                
+                //if there are still agreements attached, it isn't aloud to be deleted
+                $condition = new EqualityCondition(InternshipOrganizerAgreement :: PROPERTY_PERIOD_ID, $period_id);
+                
+                if (InternshipOrganizerDataManager::get_instance()->count_agreements($condition) > 0)
                 {
-                    //                    Event :: trigger('delete', 'period', array('target_period_id' => $period->get_id(), 'action_user_id' => $user->get_id()));
+                    $status = false;
+                    $message = 'PeriodNotDeleted-HasAgreements';
                 }
+                
+                if ($status)
+                {
+                    if (! $period->delete())
+                    {
+                        $failures ++;
+                    }
+                    else
+                    {
+                        //                    Event :: trigger('delete', 'period', array('target_period_id' => $period->get_id(), 'action_user_id' => $user->get_id()));
+                    }
+                    
+                    if ($failures)
+                    {
+                        if (count($ids) == 1)
+                        {
+                            $message = 'SelectedInternshipOrganizerPeriodNotDeleted';
+                        }
+                        else
+                        {
+                            $message = 'SelectedInternshipOrganizerPeriodsNotDeleted';
+                        }
+                    }
+                    else
+                    {
+                        if (count($ids) == 1)
+                        {
+                            $message = 'SelectedInternshipOrganizerPeriodDeleted';
+                        }
+                        else
+                        {
+                            $message = 'SelectedInternshipOrganizerPeriodsDeleted';
+                        }
+                    }
+                }
+            
             }
             
-            if ($failures)
-            {
-                if (count($ids) == 1)
-                {
-                    $message = 'SelectedInternshipOrganizerPeriodNotDeleted';
-                }
-                else
-                {
-                    $message = 'SelectedInternshipOrganizerPeriodsNotDeleted';
-                }
-            }
-            else
-            {
-                if (count($ids) == 1)
-                {
-                    $message = 'SelectedInternshipOrganizerPeriodDeleted';
-                }
-                else
-                {
-                    $message = 'SelectedInternshipOrganizerPeriodsDeleted';
-                }
-            }
-            $this->redirect(Translation :: get($message), ($failures ? true : false), array(Application :: PARAM_ACTION => InternshipOrganizerPeriodManager :: ACTION_BROWSE_PERIODS));
+            $this->redirect(Translation :: get($message), !$status, array(InternshipOrganizerPeriodManager :: PARAM_ACTION => InternshipOrganizerPeriodManager :: ACTION_BROWSE_PERIODS, DynamicTabsRenderer::PARAM_SELECTED_TAB => InternshipOrganizerPeriodManagerBrowserComponent :: TAB_SUBPERIODS));
         }
         else
         {
